Implemented single activity structure, fragment navigation, redesigned home page UI, and server choice
