+++ .idea/misc.xml
@@ -3,11 +3,13 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="..\:/Users/space/AndroidStudioProjects/SearchViewApp/app/src/main/res/layout/activity_main.xml" value="0.3625" />
+        <entry key="..\:/Users/space/AndroidStudioProjects/SearchViewApp/app/src/main/res/layout/activity_main.xml" value="0.1" />
         <entry key="..\:/Users/space/AndroidStudioProjects/SearchViewApp/app/src/main/res/layout/activity_searchable.xml" value="0.22" />
         <entry key="..\:/Users/space/AndroidStudioProjects/SearchViewApp/app/src/main/res/layout/card_view.xml" value="0.25" />
         <entry key="..\:/Users/space/AndroidStudioProjects/SearchViewApp/app/src/main/res/layout/content_main.xml" value="0.1867007672634271" />
         <entry key="..\:/Users/space/AndroidStudioProjects/SearchViewApp/app/src/main/res/layout/fragment_first.xml" value="0.21" />
+        <entry key="..\:/Users/space/AndroidStudioProjects/SearchViewApp/app/src/main/res/layout/fragment_main.xml" value="0.33" />
+        <entry key="..\:/Users/space/AndroidStudioProjects/SearchViewApp/app/src/main/res/layout/fragment_search_results.xml" value="0.34479166666666666" />
         <entry key="..\:/Users/space/AndroidStudioProjects/SearchViewApp/app/src/main/res/layout/list_item.xml" value="0.34479166666666666" />
         <entry key="..\:/Users/space/AndroidStudioProjects/SearchViewApp/app/src/main/res/layout/search.xml" value="0.33541666666666664" />
         <entry key="..\:/Users/space/AndroidStudioProjects/SearchViewApp/app/src/main/res/menu/menu_main.xml" value="0.34479166666666666" />
