+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.example.searchviewapp">
+    package="com.isaacmanu.tftstatsapp">
 
     <uses-permission android:name="android.permission.INTERNET" />
 
@@ -16,32 +16,17 @@
         android:theme="@style/Theme.SearchViewApp"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.isaacmanu.tftstatsapp.MainActivity"
             android:exported="true"
             android:windowSoftInputMode="stateVisible"
             android:label="@string/app_name"
             android:theme="@style/Theme.SearchViewApp">
-            <meta-data
-                android:name="android.app.default_searchable"
-                android:value=".SearchableActivity"/>
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
 
-        </activity>
-        <activity
-            android:name=".SearchableActivity"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.SEARCH" />
-
-                <category android:name="android.intent.category.DEFAULT"/>
-            </intent-filter>
-            <meta-data
-                android:name="android.app.searchable"
-                android:resource="@xml/searchable" />
         </activity>
     </application>
