--- app/src/main/java/com/example/searchviewapp/MainActivity.kt
@@ -1,48 +1,0 @@
-package com.example.searchviewapp
-
-import android.app.SearchManager
-import android.content.ComponentName
-import android.content.Context
-import android.os.Bundle
-import com.google.android.material.snackbar.Snackbar
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.WindowCompat
-import androidx.navigation.findNavController
-import androidx.navigation.ui.AppBarConfiguration
-import androidx.navigation.ui.navigateUp
-import androidx.navigation.ui.setupActionBarWithNavController
-import android.view.Menu
-import android.view.MenuItem
-import androidx.appcompat.widget.SearchView
-import com.example.searchviewapp.databinding.ActivityMainBinding
-
-class MainActivity : AppCompatActivity() {
-
-
-    private lateinit var binding: ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        WindowCompat.setDecorFitsSystemWindows(window, false)
-        super.onCreate(savedInstanceState)
-
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        setSupportActionBar(binding.toolbar)
-
-
-        //Setup for searchview
-        val searchManager = getSystemService(Context.SEARCH_SERVICE) as SearchManager
-        val searchView =  binding.searchView    //findViewById<SearchView>(R.id.searchView)
-        val searchableInfo = searchManager.getSearchableInfo(componentName)
-        searchView.setSearchableInfo(searchableInfo)
-
-        //Brings the searchview into focus on app start and on navigation back to activity
-        binding.searchView.requestFocus()
-
-
-
-
-
-    }
-}
