--- app/src/main/java/com/example/searchviewapp/network/model/Companion.kt
@@ -1,7 +1,0 @@
-package com.example.searchviewapp.network.model
-
-data class Companion(
-    val content_ID: String,
-    val skin_ID: Int,
-    val species: String
-)
