--- app/src/main/java/com/example/searchviewapp/network/model/MatchData.kt
@@ -1,6 +1,0 @@
-package com.example.searchviewapp.network.model
-
-data class MatchData(
-    val info: Info,
-    val metadata: Metadata
-)
