--- app/src/main/java/com/example/searchviewapp/network/model/MatchHistory.kt
@@ -1,7 +1,0 @@
-package com.example.searchviewapp.network.model
-
-import com.squareup.moshi.Json
-
-data class MatchHistory(
-    val matchId: String
-)
