--- app/src/main/java/com/example/searchviewapp/network/model/Metadata.kt
@@ -1,7 +1,0 @@
-package com.example.searchviewapp.network.model
-
-data class Metadata(
-    val data_version: String,
-    val match_id: String,
-    val participants: List<String>
-)
