--- app/src/main/java/com/example/searchviewapp/network/model/RecyclerViewItem.kt
@@ -1,6 +1,0 @@
-package com.example.searchviewapp.network.model
-
-data class RecyclerViewItem(
-    val puuid: String,
-    val matchData: MatchData
-)
