--- app/src/main/java/com/example/searchviewapp/network/model/SummonerData.kt
@@ -1,11 +1,0 @@
-package com.example.searchviewapp.network.model
-
-data class SummonerData(
-    val accountId: String,
-    val id: String,
-    val name: String,
-    val profileIconId: Int,
-    val puuid: String,
-    val revisionDate: Long,
-    val summonerLevel: Int
-)
