--- app/src/main/java/com/example/searchviewapp/network/model/Trait.kt
@@ -1,9 +1,0 @@
-package com.example.searchviewapp.network.model
-
-data class Trait(
-    val name: String,
-    val num_units: Int,
-    val style: Int,
-    val tier_current: Int,
-    val tier_total: Int
-)
