--- app/src/main/java/com/example/searchviewapp/network/model/Unit.kt
@@ -1,10 +1,0 @@
-package com.example.searchviewapp.network.model
-
-data class Unit(
-    val character_id: String,
-    val itemNames: List<String>,
-    val items: List<Int>,
-    val name: String,
-    val rarity: Int,
-    val tier: Int
-)
