+++ app/src/main/java/com/isaacmanu/tftstatsapp/network/model/Companion.kt
@@ -1,0 +1,7 @@
+package com.isaacmanu.tftstatsapp.network.model
+
+data class Companion(
+    val content_ID: String,
+    val skin_ID: Int,
+    val species: String
+)
