+++ app/src/main/java/com/isaacmanu/tftstatsapp/network/model/MatchData.kt
@@ -1,0 +1,6 @@
+package com.isaacmanu.tftstatsapp.network.model
+
+data class MatchData(
+    val info: Info,
+    val metadata: Metadata
+)
