+++ app/src/main/java/com/isaacmanu/tftstatsapp/network/model/MatchHistory.kt
@@ -1,0 +1,5 @@
+package com.isaacmanu.tftstatsapp.network.model
+
+data class MatchHistory(
+    val matchId: String
+)
