+++ app/src/main/java/com/isaacmanu/tftstatsapp/network/model/Metadata.kt
@@ -1,0 +1,7 @@
+package com.isaacmanu.tftstatsapp.network.model
+
+data class Metadata(
+    val data_version: String,
+    val match_id: String,
+    val participants: List<String>
+)
