+++ app/src/main/java/com/isaacmanu/tftstatsapp/network/model/Participant.kt
@@ -1,0 +1,16 @@
+package com.isaacmanu.tftstatsapp.network.model
+
+data class Participant(
+    val augments: List<String>,
+    val `companion`: Companion,
+    val gold_left: Int,
+    val last_round: Int,
+    val level: Int,
+    val placement: Int,
+    val players_eliminated: Int,
+    val puuid: String,
+    val time_eliminated: Double,
+    val total_damage_to_players: Int,
+    val traits: List<Trait>,
+    val units: List<Unit>
+)
