+++ app/src/main/java/com/isaacmanu/tftstatsapp/network/model/RankedData.kt
@@ -1,0 +1,17 @@
+package com.isaacmanu.tftstatsapp.network.model
+
+data class RankedData(
+    val freshBlood: Boolean,
+    val hotStreak: Boolean,
+    val inactive: Boolean,
+    val leagueId: String,
+    val leaguePoints: Int,
+    val losses: Int,
+    val queueType: String,
+    val rank: String,
+    val summonerId: String,
+    val summonerName: String,
+    val tier: String,
+    val veteran: Boolean,
+    val wins: Int
+)
