+++ app/src/main/java/com/isaacmanu/tftstatsapp/network/model/RecyclerViewItem.kt
@@ -1,0 +1,6 @@
+package com.isaacmanu.tftstatsapp.network.model
+
+data class RecyclerViewItem(
+    val puuid: String,
+    val matchData: MatchData
+)
