+++ app/src/main/java/com/isaacmanu/tftstatsapp/network/model/SummonerData.kt
@@ -1,0 +1,11 @@
+package com.isaacmanu.tftstatsapp.network.model
+
+data class SummonerData(
+    val accountId: String,
+    val id: String,
+    val name: String,
+    val profileIconId: Int,
+    val puuid: String,
+    val revisionDate: Long,
+    val summonerLevel: Int
+)
