+++ app/src/main/java/com/isaacmanu/tftstatsapp/network/model/Trait.kt
@@ -1,0 +1,9 @@
+package com.isaacmanu.tftstatsapp.network.model
+
+data class Trait(
+    val name: String,
+    val num_units: Int,
+    val style: Int,
+    val tier_current: Int,
+    val tier_total: Int
+)
