+++ app/src/main/java/com/isaacmanu/tftstatsapp/network/model/Unit.kt
@@ -1,0 +1,10 @@
+package com.isaacmanu.tftstatsapp.network.model
+
+data class Unit(
+    val character_id: String,
+    val itemNames: List<String>,
+    val items: List<Int>,
+    val name: String,
+    val rarity: Int,
+    val tier: Int
+)
