+++ app/src/main/res/layout/activity_main.xml
@@ -24,29 +24,19 @@
     </com.google.android.material.appbar.AppBarLayout>
 
 
-    <com.google.android.material.button.MaterialButton
-        android:id="@+id/server_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toStartOf="@id/searchView"
-        app:layout_constraintTop_toTopOf="@id/searchView"
-        app:layout_constraintTop_toBottomOf="@id/searchView"
-        android:text="EUW"
-        />
-
-
-    <androidx.appcompat.widget.SearchView
-        android:id="@+id/searchView"
-        android:layout_width="250dp"
-        android:layout_height="wrap_content"
-        app:iconifiedByDefault="false"
-        android:inputType="text"
-        app:queryHint="Enter Username"
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/nav_host_fragment"
+        android:name="androidx.navigation.fragment.NavHostFragment"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/server_button"
-        app:layout_constraintTop_toTopOf="parent" />
+
+        app:defaultNavHost="true"
+        app:navGraph="@navigation/nav_graph" />
+
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
