+++ app/src/main/res/navigation/nav_graph.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_graph"
+    app:startDestination="@id/mainFragment">
+    <fragment
+        android:id="@+id/mainFragment"
+        android:name="com.isaacmanu.tftstatsapp.MainFragment"
+        android:label="TFT Stats App"
+        tools:layout="@layout/fragment_main">
+        <action
+            android:id="@+id/action_mainFragment_to_searchResultsFragment"
+            app:destination="@id/searchResultsFragment" />
+    </fragment>
+    <fragment
+        android:id="@+id/searchResultsFragment"
+        android:name="com.isaacmanu.tftstatsapp.SearchResultsFragment"
+        android:label="Match History"
+        tools:layout="@layout/fragment_search_results" >
+        <argument
+            android:name="username"
+            app:argType="string" />
+        <argument
+            android:name="server"
+            app:argType="string" />
+    </fragment>
+</navigation>
