+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <resources>
-    <string name="app_name">SearchViewApp</string>
+    <string name="app_name">TFT Stats</string>
     <string name="action_settings">Settings</string>
-    <string name="search_hint">Enter Username</string>
+    <string name="search_hint">Username</string>
     <!-- Strings used for fragments for navigation -->
     <string name="first_fragment_label">First Fragment</string>
     <string name="second_fragment_label">Second Fragment</string>
@@ -11,4 +11,6 @@
     <string name="level">Level: %1$s</string>
     <string name="rank">%1$s %2$s, %3$d LP</string>
     <string name="user_profile_picture">User Profile Picture</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
